Moved SU* test classes to corresponding math package.
