+++ pixi/src/test/java/org/openpixi/pixi/math/SU2MatrixTest.java
@@ -1,4 +1,4 @@
-package org.openpixi.pixi.physics.grid;
+package org.openpixi.pixi.math;
 
 
 import org.apache.commons.math3.Field;
